chore(core): Move diffValue.ts and getJsonStream.ts into the events folder
