+++ packages/sanity/src/core/store/translog/getTransactionsLogs.test.ts
@@ -1,10 +1,10 @@
 import {type SanityClient} from '@sanity/client'
 import {beforeEach, describe, expect, it, type Mock, vi} from 'vitest'
 
-import {getJsonStream} from '../history/history/getJsonStream'
+import {getJsonStream} from '../events/getJsonStream'
 import {getTransactionsLogs} from './getTransactionsLogs'
 
-vi.mock('../history/history/getJsonStream', () => ({
+vi.mock('../events/getJsonStream', () => ({
   getJsonStream: vi.fn(),
 }))
